ReadableBuffer cleanup and changes - Make first computed to save some room - Move Equals and StartsWith to extensions
